requirements.txt
Django==5.0.1
PyMuPDFb==1.23.9
python-dotenv==1.0.0
langchain-core==0.1.11
openai==1.8.0
tiktoken
